pyproject.toml
[project]
name = "paralympics"
version = "1.0.0"
description = "Paralympics testing activities"
dependencies = [
    "flask",
    "Flask-SQLAlchemy",
    "Flask-Marshmallow",
    "marshmallow-sqlalchemy",
    "pandas",
    "selenium",
    "pytest"
]

[build-system]
requires = [
    "setuptools>=65.0",
    "setuptools-scm[toml]>=6.2.3",
]
build-backend = "setuptools.build_meta"

# Optional, these configure Pytest behaviour
[tool.pytest.ini_options]
minversion = "2.7"
addopts = "-rfEX -p pytester --strict-markers"
python_files = ["test_*.py", "*_test.py", "test/*.py"]
python_classes = ["Test"]
python_functions = ["test"]
testpaths = ["test"]
xfail_strict = true
filterwarnings = [
    "error",
    # pytest's own futurewarnings
    "ignore::pytest.PytestExperimentalApiWarning",
]
